adding new clims
